+++ menu.v
@@ -67,8 +67,8 @@ fn (mut m Menu) draw() {
 		return
 	}
 	gg := m.ui.gg
-	gg.draw_rect(m.x, m.y, 150, m.items.len * menu_height, menu_color)
-	gg.draw_empty_rect(m.x, m.y, 150, m.items.len * menu_height, menu_border_color)
+	gg.draw_rect(f32(m.x), f32(m.y), 150, f32(m.items.len * menu_height), menu_color)
+	gg.draw_empty_rect(f32(m.x), f32(m.y), 150, f32(m.items.len * menu_height), menu_border_color)
 	for i, item in m.items {
 		m.ui.ft.draw_text_def(m.x + 10, m.y + i * menu_height  +10, item.text)
 	}
